+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+* Fixed mounting 2 or more drowneds at the same time
+* Fixed stuck on drowned on respawn
+
 ## 3.3.2
 * Increased default despawn distance (96 -> 112) 
 * Minor updates to config comments